Scripts: install_all_pre-commit will not replace existing pre-commit hook

+++ scripts/install_all_pre-commit.sh
@@ -13,7 +13,10 @@ for one_project in "${ALL_PROJECTS[@]}"; do
     if [ -d "$one_project/.git" ]; then
 
         pushd "$one_project"
-        cp "$PROJECTS_DIR/config-sh/git_hooks/author-guard.sh" ".git/hooks/pre-commit"
+
+        if [ ! -f ".git/hooks/pre-commit" ]; then
+            cp "$PROJECTS_DIR/config-sh/git_hooks/author-guard.sh" ".git/hooks/pre-commit"
+        fi
 
         ln --symbolic --force --no-target-directory "$PROJECTS_DIR/config-sh/git_hooks/push-guard.sh" ".git/hooks/pre-push"
